+++ src/main/java/kitchenpos/menu/domain/MenuProductDao.java
@@ -1,6 +1,6 @@
-package kitchenpos.dao;
+package kitchenpos.menu.domain;
 
-import kitchenpos.domain.MenuProduct;
+import kitchenpos.menu.domain.MenuProduct;
 
 import java.util.List;
 import java.util.Optional;
